requirements.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
blessed==1.19.1
Django==4.1.4
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-picklefield==3.1
django-q==1.3.9
jinxed==1.2.0
Pillow==9.3.0
python-dateutil==2.8.2
redis==3.5.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
wcwidth==0.2.5
